requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
social-auth-app-django==5.4.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
pyotp==2.9.0
python-decouple==3.8
django-environ==0.11.2